+++ apps/dotcom/client/src/tla/components/TlaAppMenuGroup/TlaAppMenuGroup.tsx
@@ -1,21 +1,7 @@
-import {
-	ColorSchemeMenu,
-	LanguageMenu,
-	TldrawUiMenuGroup,
-	TldrawUiMenuSubmenu,
-	ToggleDebugModeItem,
-	ToggleDynamicSizeModeItem,
-	ToggleEdgeScrollingItem,
-	ToggleFocusModeItem,
-	ToggleGridItem,
-	TogglePasteAtCursorItem,
-	ToggleReduceMotionItem,
-	ToggleSnapModeItem,
-	ToggleToolLockItem,
-	ToggleWrapModeItem,
-} from 'tldraw'
+import { ColorSchemeMenu, LanguageMenu, TldrawUiMenuGroup, TldrawUiMenuSubmenu } from 'tldraw'
 import { Links } from '../../../components/Links'
 import { defineMessages, useMsg } from '../../utils/i18n'
+import { TlaDebugMenuGroup } from '../TlaDebugMenuGroup'
 
 const messages = defineMessages({
 	help: { defaultMessage: 'Help' },
@@ -27,6 +13,7 @@ export function TlaAppMenuGroup() {
 			<HelpSubMenu />
 			<ColorThemeSubmenu />
 			<LanguageMenu />
+			<TlaDebugMenuGroup />
 		</TldrawUiMenuGroup>
 	)
 }
@@ -43,28 +30,6 @@ export function TlaAppMenuGroupLazyFlipped() {
 
 function ColorThemeSubmenu() {
 	return <ColorSchemeMenu />
-}
-
-function _PreferencesSubmenu() {
-	return (
-		<TldrawUiMenuSubmenu id="preferences" label="menu.preferences">
-			<TldrawUiMenuGroup id="preferences-actions">
-				<ToggleSnapModeItem />
-				<ToggleToolLockItem />
-				<ToggleGridItem />
-				<ToggleWrapModeItem />
-				<ToggleFocusModeItem />
-				<ToggleEdgeScrollingItem />
-				<ToggleReduceMotionItem />
-				<ToggleDynamicSizeModeItem />
-				<TogglePasteAtCursorItem />
-				<ToggleDebugModeItem />
-			</TldrawUiMenuGroup>
-			<TldrawUiMenuGroup id="color-scheme">
-				<ColorSchemeMenu />
-			</TldrawUiMenuGroup>
-		</TldrawUiMenuSubmenu>
-	)
 }
 
 function HelpSubMenu() {