+++ apps/dotcom/client/src/tla/hooks/useTldrFileDrop.ts
@@ -1,17 +1,22 @@
 import { useAuth } from '@clerk/clerk-react'
 import { DragEvent, useCallback, useState } from 'react'
+import { useNavigate } from 'react-router-dom'
 import { Editor, TLStoreSnapshot, parseTldrawJsonFile, tlmenus, useToasts } from 'tldraw'
 import { globalEditor } from '../../utils/globalEditor'
 import { defineMessages, useIntl } from '../utils/i18n'
+import { getFilePath } from '../utils/urls'
 import { useApp } from './useAppState'
 
 const messages = defineMessages({
-	uploading: { defaultMessage: 'Uploading .tldr {count, plural, one {# file} other {# files}}…' },
-	adding: { defaultMessage: 'Added .tldr {count, plural, one {# file} other {# files}}' },
+	uploading: {
+		defaultMessage: 'Uploading {count} .tldr {count, plural, one {file} other {files}}…',
+	},
+	adding: { defaultMessage: 'Added {count} .tldr {count, plural, one {file} other {files}}.' },
 })
 
 export function useTldrFileDrop() {
 	const app = useApp()
+	const navigate = useNavigate()
 
 	const [isDraggingOver, setIsDraggingOver] = useState(false)
 
@@ -52,7 +57,7 @@ export function useTldrFileDrop() {
 					title: uploadingTitle,
 				})
 
-				await app.createFilesFromTldrFiles(snapshots, token)
+				const results = await app.createFilesFromTldrFiles(snapshots, token)
 
 				removeToast(id)
 				addToast({
@@ -60,9 +65,12 @@ export function useTldrFileDrop() {
 					title: addedTitle,
 					keepOpen: true,
 				})
+				if (results.slugs.length > 0) {
+					navigate(getFilePath(results.slugs[0]), { state: { mode: 'create' } })
+				}
 			}
 		},
-		[app, addToast, removeToast, auth, intl]
+		[app, addToast, removeToast, auth, intl, navigate]
 	)
 
 	const onDragOver = useCallback((e: DragEvent) => {